fix: Changed location of index.html and fixed broken links for images and pages
To make my project available to GitHub pages, the index.html must be located in the root of the Repo
directtory. The index.html was in a "src" directory, so a change was required. Image links and page
links became broken and had to fix those

starting # with '#' will be ignored, and an empty message aborts the commit.